Refactor: extract markers and layers into subdirectories
- Move StopMarkers/StopMarker from explore-map.tsx → markers/stop-marker.tsx
- Move vehicle-map-marker.tsx → markers/vehicle-map-marker.tsx
- Extract StopsGeojsonLayer from trip-map.tsx → layers/stops-layer.tsx (renamed StopsLayer)
- Extract ShapesGeojsonLayer from trip-map.tsx → layers/shape-layer.tsx (renamed ShapeLayer)
- Extract LiveVehiclesLayer from trip-map.tsx → layers/vehicles-layer.tsx (renamed VehiclesLayer)
- Move TripMapStopInfo type to shared types.ts to break circular dependency
- Clean up imports in explore-map.tsx and trip-map.tsx
- Delete old vehicle-map-marker.tsx from maps root
